Move CustomHttpResponseStatus codes to styx-proxy module.
