+++ components/proxy/src/test/java/com/hotels/styx/proxy/HttpErrorStatusMetricsTest.java
@@ -56,9 +56,9 @@ import static com.hotels.styx.api.HttpResponseStatus.REQUEST_URI_TOO_LONG;
 import static com.hotels.styx.api.HttpResponseStatus.UNAUTHORIZED;
 import static com.hotels.styx.api.HttpResponseStatus.UNSUPPORTED_MEDIA_TYPE;
 import static com.hotels.styx.proxy.HttpErrorStatusMetrics.formattedExceptionName;
-import static com.hotels.styx.api.CustomHttpResponseStatus.ORIGIN_CONNECTION_REFUSED;
-import static com.hotels.styx.api.CustomHttpResponseStatus.ORIGIN_CONNECTION_TIMED_OUT;
-import static com.hotels.styx.api.CustomHttpResponseStatus.ORIGIN_SERVER_TIMED_OUT;
+import static com.hotels.styx.CustomHttpResponseStatus.ORIGIN_CONNECTION_REFUSED;
+import static com.hotels.styx.CustomHttpResponseStatus.ORIGIN_CONNECTION_TIMED_OUT;
+import static com.hotels.styx.CustomHttpResponseStatus.ORIGIN_SERVER_TIMED_OUT;
 import static java.lang.System.arraycopy;
 import static java.util.stream.Collectors.toList;
 import static org.hamcrest.MatcherAssert.assertThat;
